Weather/requirements.txt
python-dateutil
numpy
streamlit
pandas
plotly
python-dotenv
py-clob-client
